--- core/src/main/java/tech/underoaks/coldcase/data/tileContent/TestItem02.java
@@ -1,11 +1,0 @@
-package tech.underoaks.coldcase.data.tileContent;
-
-import com.badlogic.gdx.graphics.Texture;
-
-public class TestItem02 extends ItemObject{
-    private static final Texture texture = new Texture("./isometric tileset/separated images/TEST_ITEM02.png");
-
-    public TestItem02() {
-        super(texture);
-    }
-}
