--- core/src/main/java/tech/underoaks/coldcase/enums/VisibilityStates.java
@@ -1,19 +1,0 @@
-package tech.underoaks.coldcase.enums;
-
-/**
- * Enum representing the visibility state of a TileContent.
- */
-public enum VisibilityStates {
-    /**
-     * Visible only to Player One.
-     */
-    PLAYER_ONE_ONLY,
-    /**
-     * Visible only to Player Two.
-     */
-    PLAYER_TWO_ONLY,
-    /**
-     * Visible to all players.
-     */
-    TRANSCENDENT
-}
