+++ core/src/main/java/tech/underoaks/coldcase/state/tileContent/Player.java
@@ -1,0 +1,54 @@
+package tech.underoaks.coldcase.state.tileContent;
+
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.math.Vector2;
+import tech.underoaks.coldcase.state.updates.GameStateUpdateException;
+import tech.underoaks.coldcase.state.InteractionChain;
+import tech.underoaks.coldcase.state.updates.MoveUpdate;
+import tech.underoaks.coldcase.state.Map;
+import tech.underoaks.coldcase.state.tiles.Tile;
+import tech.underoaks.coldcase.game.Direction;
+
+public class Player extends TileContent {
+
+    private static final Texture texture = new Texture("./isometric tileset/placholder_player.png");
+
+    public Player() {
+        super(texture, true, false);
+    }
+
+    @Override
+    public boolean action(InteractionChain chain, Vector2 tilePosition, Direction actionDirection) throws GameStateUpdateException {
+
+        int childIndex = chain.getSnapshot().getSnapshotMap().getChildIndex(tilePosition, this);
+        Vector2 targetPosition = switch (actionDirection) {
+            case NORTH -> tilePosition.cpy().add(0, -1);
+            case SOUTH -> tilePosition.cpy().add(0, 1);
+            case EAST -> tilePosition.cpy().add(1, 0);
+            case WEST -> tilePosition.cpy().add(-1, 0);
+        };
+
+        Map snapshotMap = chain.getSnapshot().getSnapshotMap();
+
+        //validation checks: if the target tile is out of bounds
+        if (snapshotMap.isOutOfBounds(targetPosition)) {
+            return false;
+        }
+
+        //validation checks: if the target tile is not passable
+        Tile targetTile = snapshotMap.getTile(targetPosition);
+        TileContent topTargetContent = targetTile.topTileContent();
+        if (topTargetContent != null && !topTargetContent.isPlayerPassable()) {
+            return false;
+        }
+
+        // adding the validated movement to the chain
+        chain.addGameStateUpdate(new MoveUpdate(tilePosition, childIndex, targetPosition));
+        return true;
+    }
+
+    @Override
+    public boolean update(InteractionChain chain, Vector2 tilePosition) throws GameStateUpdateException {
+        return false;
+    }
+}
